配置変更、花火追加
