+++ src/main/java/cn/nukkit/Server.java
@@ -60,14 +60,10 @@ import cn.nukkit.command.SimpleCommandMap;
 import cn.nukkit.entity.Attribute;
 import cn.nukkit.entity.Entity;
 import cn.nukkit.entity.EntityHuman;
-import cn.nukkit.entity.boss.EntityElderGuardian;
-import cn.nukkit.entity.boss.EntityEnderDragon;
-import cn.nukkit.entity.boss.EntityWither;
 import cn.nukkit.entity.data.Skin;
 import cn.nukkit.entity.item.EntityArmorStand;
 import cn.nukkit.entity.item.EntityBoat;
 import cn.nukkit.entity.item.EntityEnderCrystal;
-import cn.nukkit.entity.item.EntityExpBottle;
 import cn.nukkit.entity.item.EntityFallingBlock;
 import cn.nukkit.entity.item.EntityItem;
 import cn.nukkit.entity.item.EntityMinecartChest;
@@ -75,12 +71,13 @@ import cn.nukkit.entity.item.EntityMinecartEmpty;
 import cn.nukkit.entity.item.EntityMinecartHopper;
 import cn.nukkit.entity.item.EntityMinecartTNT;
 import cn.nukkit.entity.item.EntityPainting;
-import cn.nukkit.entity.item.EntityPotion;
 import cn.nukkit.entity.item.EntityPrimedTNT;
 import cn.nukkit.entity.item.EntityXPOrb;
 import cn.nukkit.entity.mob.EntityBlaze;
 import cn.nukkit.entity.mob.EntityCaveSpider;
 import cn.nukkit.entity.mob.EntityCreeper;
+import cn.nukkit.entity.mob.EntityElderGuardian;
+import cn.nukkit.entity.mob.EntityEnderDragon;
 import cn.nukkit.entity.mob.EntityEnderman;
 import cn.nukkit.entity.mob.EntityEndermite;
 import cn.nukkit.entity.mob.EntityEvoker;
@@ -97,6 +94,7 @@ import cn.nukkit.entity.mob.EntityStray;
 import cn.nukkit.entity.mob.EntityVex;
 import cn.nukkit.entity.mob.EntityVindicator;
 import cn.nukkit.entity.mob.EntityWitch;
+import cn.nukkit.entity.mob.EntityWither;
 import cn.nukkit.entity.mob.EntityWitherSkeleton;
 import cn.nukkit.entity.mob.EntityZombie;
 import cn.nukkit.entity.mob.EntityZombiePigman;
@@ -124,7 +122,9 @@ import cn.nukkit.entity.passive.EntityWolf;
 import cn.nukkit.entity.passive.EntityZombieHorse;
 import cn.nukkit.entity.projectile.EntityArrow;
 import cn.nukkit.entity.projectile.EntityEnderPearl;
+import cn.nukkit.entity.projectile.EntityExpBottle;
 import cn.nukkit.entity.projectile.EntityFishingHook;
+import cn.nukkit.entity.projectile.EntityPotion;
 import cn.nukkit.entity.projectile.EntitySnowball;
 import cn.nukkit.entity.weather.EntityLightning;
 import cn.nukkit.event.HandlerList;
