+++ src/main/java/cn/nukkit/entity/mob/EntityWither.java
@@ -1,4 +1,4 @@
-package cn.nukkit.entity.boss;
+package cn.nukkit.entity.mob;
 
 import cn.nukkit.Player;
 import cn.nukkit.item.Item;
