+++ src/main/java/cn/nukkit/entity/projectile/EntityEnderPearl.java
@@ -2,85 +2,98 @@ package cn.nukkit.entity.projectile;
 
 import cn.nukkit.Player;
 import cn.nukkit.entity.Entity;
-import cn.nukkit.event.player.PlayerTeleportEvent.TeleportCause;
+import cn.nukkit.event.entity.EntityDamageEvent;
+import cn.nukkit.event.entity.EntityDamageEvent.DamageCause;
 import cn.nukkit.level.format.FullChunk;
-import cn.nukkit.level.sound.EndermanTeleportSound;
-import cn.nukkit.math.NukkitMath;
-import cn.nukkit.math.Vector3;
 import cn.nukkit.nbt.tag.CompoundTag;
 import cn.nukkit.network.protocol.AddEntityPacket;
 
-public class EntityEnderPearl extends EntityProjectile {
-    public static final int NETWORK_ID = 87;
+public class EntityEnderPearl extends EntityProjectile{
 
-    @Override
-    public int getNetworkId() {
-        return NETWORK_ID;
-    }
+	public static final int NETWORK_ID = 87;
 
-    @Override
-    public float getWidth() {
-        return 0.25f;
-    }
+	public float width = 0.25f;
+	public float length = 0.25f;
+	public float height = 0.25f;
 
-    @Override
-    public float getLength() {
-        return 0.25f;
-    }
+	protected float gravity = 0.03f;
+	protected float drag = 0.01f;
 
-    @Override
-    public float getHeight() {
-        return 0.25f;
-    }
+	private boolean hasTeleportedShooter = false;
 
-    @Override
-    protected float getGravity() {
-        return 0.03f;
-    }
+	public EntityEnderPearl(FullChunk chunk, CompoundTag nbt, Entity shootingEntity) {
+		super(chunk, nbt, shootingEntity);
+	}
 
-    @Override
-    protected float getDrag() {
-        return 0.01f;
-    }
+	public EntityEnderPearl(FullChunk chunk, CompoundTag nbt){
+		this(chunk, nbt, null);
+	}
 
-    public EntityEnderPearl(FullChunk chunk, CompoundTag nbt) {
-        this(chunk, nbt, null);
-    }
+	@Override
+	public int getNetworkId() {
+		return NETWORK_ID;
+	}
 
-    public EntityEnderPearl(FullChunk chunk, CompoundTag nbt, Entity shootingEntity) {
-        super(chunk, nbt, shootingEntity);
-    }
+	@Override
+	public float getWidth(){
+		return width;
+	}
 
-    @Override
-    public boolean onUpdate(int currentTick) {
-        if (this.closed) {
-            return false;
-        }
+	@Override
+	public float getLength(){
+		return length;
+	}
 
-        this.timing.startTiming();
+	@Override
+	public float getHeight(){
+		return height;
+	}
 
-        boolean hasUpdate = super.onUpdate(currentTick);
+	@Override
+	protected float getGravity(){
+		return gravity;
+	}
 
-        if (this.isCollided && this.shootingEntity instanceof Player) {
-            this.shootingEntity.teleport(new Vector3(NukkitMath.floorDouble(this.x) + 0.5, this.y, NukkitMath.floorDouble(this.z) + 0.5), TeleportCause.ENDER_PEARL);
-            if ((((Player) this.shootingEntity).getGamemode() & 0x01) == 0) this.shootingEntity.attack(5);
-            this.level.addSound(new EndermanTeleportSound(this));
-        }
+	@Override
+	protected float getDrag(){
+		return drag;
+	}
+
+	public void teleportShooter(){
+		if(!this.hasTeleportedShooter){
+			this.hasTeleportedShooter = true;
+			if(this.shootingEntity instanceof Player && this.y > 0){
+				this.shootingEntity.attack(new EntityDamageEvent(this.shootingEntity, DamageCause.FALL, 5));
+				this.shootingEntity.teleport(this.getLocation());
+			}
+			this.kill();
+		}
+	}
+
+	@Override
+	public boolean onUpdate(int currentTick){
+		if(this.closed){
+			return false;
+		}
+
+		this.timing.startTiming();
+
+		boolean hasUpdate = super.onUpdate(currentTick);
 
         if (this.age > 1200 || this.isCollided) {
-            this.kill();
+            this.teleportShooter();
             hasUpdate = true;
         }
 
         this.timing.stopTiming();
 
         return hasUpdate;
-    }
+	}
 
-    @Override
-    public void spawnTo(Player player) {
+	@Override
+	public void spawnTo(Player player){
         AddEntityPacket pk = new AddEntityPacket();
-        pk.type = NETWORK_ID;
+        pk.type = EntityEnderPearl.NETWORK_ID;
         pk.entityUniqueId = this.getId();
         pk.entityRuntimeId = this.getId();
         pk.x = (float) this.x;
@@ -93,5 +106,5 @@ public class EntityEnderPearl extends EntityProjectile {
         player.dataPacket(pk);
 
         super.spawnTo(player);
-    }
+	}
 }
