+++ src/main/java/cn/nukkit/entity/projectile/EntityFireworkRocket.java
@@ -1,0 +1,74 @@
+package cn.nukkit.entity.projectile;
+
+import cn.nukkit.Player;
+import cn.nukkit.entity.Entity;
+import cn.nukkit.level.format.FullChunk;
+import cn.nukkit.nbt.tag.CompoundTag;
+import cn.nukkit.network.protocol.AddEntityPacket;
+
+/**
+ * @author Megapix96
+ */
+public class EntityFireworkRocket extends EntityProjectile {
+
+    public static final int NETWORK_ID = 72;
+
+    @Override
+    public int getNetworkId() {
+        return NETWORK_ID;
+    }
+
+    @Override
+    protected float getGravity() {
+        return 0.1f;
+    }
+
+    @Override
+    protected float getDrag() {
+        return 0.01f;
+    }
+
+    public EntityFireworkRocket(FullChunk chunk, CompoundTag nbt) {
+        this(chunk, nbt, null);
+    }
+
+    public EntityFireworkRocket(FullChunk chunk, CompoundTag nbt, Entity shootingEntity) {
+        super(chunk, nbt, shootingEntity);
+    }
+
+    @Override
+    public boolean onUpdate(int currentTick) {
+        if (this.closed) {
+            return false;
+        }
+
+        this.timing.startTiming();
+
+        int tickDiff = currentTick - this.lastUpdate;
+        boolean hasUpdate = super.onUpdate(currentTick);
+
+        //TODO
+
+        this.timing.stopTiming();
+
+        return hasUpdate;
+    }
+
+    @Override
+    public void spawnTo(Player player) {
+        AddEntityPacket pk = new AddEntityPacket();
+        pk.type = EntityFireworkRocket.NETWORK_ID;
+        pk.entityUniqueId = this.getId();
+        pk.entityRuntimeId = this.getId();
+        pk.x = (float) this.x;
+        pk.y = (float) this.y;
+        pk.z = (float) this.z;
+        pk.speedX = (float) this.motionX;
+        pk.speedY = (float) this.motionY;
+        pk.speedZ = (float) this.motionZ;
+        pk.metadata = this.dataProperties;
+        player.dataPacket(pk);
+
+        super.spawnTo(player);
+    }
+}
