+++ src/main/java/cn/nukkit/entity/projectile/EntityPotion.java
@@ -1,8 +1,7 @@
-package cn.nukkit.entity.item;
+package cn.nukkit.entity.projectile;
 
 import cn.nukkit.Player;
 import cn.nukkit.entity.Entity;
-import cn.nukkit.entity.projectile.EntityProjectile;
 import cn.nukkit.event.potion.PotionCollideEvent;
 import cn.nukkit.level.format.FullChunk;
 import cn.nukkit.level.particle.InstantSpellParticle;
